use (long long) type for pdb_entry_num

+++ qTMclust.cpp
@@ -157,7 +157,7 @@ int main(int argc, char **argv)
     cout<<"write output for TM-score "<<TMmin<<endl;
     write_TMclust_result("cluster.txt",TMclust,pdb_name_list,TMmin);
     write_matrix("TM_fast.txt",tm_fast_mat);
-    write_matrix("TM_full.txt",tm_full_mat,pdb_entry_num);
+    write_matrix("TM_full.txt",tm_full_mat,(long long)(pdb_entry_num));
     write_TMclust_ca_xyz("ca.xyz", TMclust.repr_list, 
         pdb_name_list, pdb_chain_list);
 
@@ -167,7 +167,7 @@ int main(int argc, char **argv)
     full_clustering(TMclust, pdb_name_list, pdb_file_list, pdb_chain_list, 
         tm_fast_mat, tm_full_mat, TMmin, TMmax, TMstep, "cluster.txt",
         'a', norm, MinClustSize, CacheCoor);
-    write_matrix("TM_full.txt",tm_full_mat,pdb_entry_num);
+    write_matrix("TM_full.txt",tm_full_mat,(long long)(pdb_entry_num));
 
     /* clean up */
     cout<<"finished clustering of "<<pdb_entry_num<<" chains into "
